backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
gunicorn==21.2.0
sqlalchemy>=2.0.30
psycopg[binary]
google-generativeai==0.3.2
python-dotenv==1.0.1
pgvector==0.2.4
